requirements.txt
flask-restplus==0.9.2
Flask-SQLAlchemy==2.1
alembic==0.8.8
psycopg2==2.6.2